chore: rename modules to PEP 8 snake_case
readVCF.py → read_vcf.py and DBSCAN.py → dbscan.py: module names must be
lowercase per PEP 8; CamelCase and ALLCAPS are reserved for class names.
Update all import sites (build_module, query_module, merge_vcf_module,
export_module), tests, and setup.py Cython build list.

CLI flag --DBSCAN and args.DBSCAN are unchanged — those are user-facing
argument names, not Python identifiers.
