+++ setup.py
@@ -10,8 +10,8 @@ if has_cython:
     ext_modules = cythonize([
         "svdb/build_module.py",
         "svdb/overlap_module.py",
-        "svdb/DBSCAN.py",
-        "svdb/readVCF.py",
+        "svdb/dbscan.py",
+        "svdb/read_vcf.py",
         "svdb/merge_vcf_module_cython.py",
         "svdb/query_module.py",
         "svdb/export_module.py"])
